fix the sequential data generation
