+++ packages/data/src/buildProgramJsonOutput.ts
@@ -1,0 +1,14 @@
+import { readPrograms, prependInterface, buildProgramJSONOutput } from './dataPipeline'
+// Script
+
+console.log('▶ Starting data consolidation (buildJsonOutput.ts)\n')
+
+let programs = readPrograms(true)
+
+console.log()
+
+programs = prependInterface(programs, true)
+
+console.log()
+
+buildProgramJSONOutput(programs)
